requirements.txt
asgiref==3.4.1
cffi==1.15.0
cryptography==36.0.0
Django==3.2.9
django-cors-headers==3.10.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
prettyconf==2.2.1
pycparser==2.21
PyJWT==2.3.0
pytz==2021.3
sqlparse==0.4.2
